Extended to parse 5 GB files
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1074 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/ParseOBJ.h
@@ -379,7 +379,7 @@ private:
 
 public:
 
-    void parse(const char* ptr, int len, const std::string& basePath);
+    void parse(const char* ptr, size_t len, const std::string& basePath);
 
     void parse(BinaryInput& bi, const std::string& basePath = "<AUTO>");
 };
